reorganised files and type safe env variables
