+++ src/components/forms/entry-form.tsx
@@ -1,16 +1,83 @@
 import toast from "react-hot-toast";
 import Link from "next/link";
-import SearchBar from "../fields/search-bar";
+import SearchBar from "../ui/search-bar";
 import ExerciseCard from "../cards/exercise-card";
-import SetForm from "./set-form";
 import { useRouter } from "next/router";
 import { useState } from "react";
-import { useFieldArray, useForm } from "react-hook-form";
+import {
+  useFieldArray,
+  useForm,
+  FieldArrayWithId,
+  UseFieldArrayRemove,
+  UseFormRegister
+} from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useSearch } from "@/utils/hooks/use-search";
+import { useSearch } from "@/utils/use-search";
 import { HiX } from "react-icons/hi";
-import { EntryValidator, entryValidator } from "@/utils/validators/entry";
+import { EntryValidator, entryValidator } from "@/server/shared/entry";
 import { InferQueryOutput, trpc } from "@/utils/trpc";
+
+type SetFormProps = {
+  cardio?: boolean;
+  workoutSet?: boolean;
+  set: FieldArrayWithId<EntryValidator, "sets", "id">;
+  index: number;
+  remove: UseFieldArrayRemove;
+  register: UseFormRegister<EntryValidator>;
+};
+
+const SetForm = ({ cardio, set, index, remove, register }: SetFormProps) => {
+  return (
+    <div
+      key={set.id}
+      className='py-4 px-6 bg-zinc-900 rounded-md border border-zinc-500'
+    >
+      <div className='flex flex-wrap gap-6 items-center'>
+        <div className='self-start flex gap-2 sm:flex-col'>
+          <p>Set</p>
+          <p>{index + 1}</p>
+        </div>
+
+        <div className='field'>
+          <label htmlFor={cardio ? "distance" : "reps"}>
+            {cardio ? "Distance (miles)" : "Reps"}
+          </label>
+
+          <input
+            {...register(
+              cardio ? `sets.${index}.distance` : `sets.${index}.reps`
+            )}
+            className='input-small'
+            type='number'
+            id={cardio ? "distance" : "reps"}
+            step={cardio ? "0.01" : "0"}
+          />
+        </div>
+
+        <div className='field'>
+          <label htmlFor={cardio ? "elevation" : "weight"}>
+            {cardio ? "Elevation (ft)" : "Weight (kg)"}
+          </label>
+
+          <input
+            {...register(
+              cardio ? `sets.${index}.elevation` : `sets.${index}.weight`
+            )}
+            className='input-small'
+            type='number'
+            id={cardio ? "distance" : "reps"}
+            step='0.01'
+          />
+        </div>
+
+        <button className='button-remove' onClick={() => remove(index)}>
+          <HiX className='h-5 w-5' />
+          <p className='hidden sm:inline'>Remove set</p>
+        </button>
+      </div>
+    </div>
+  );
+};
 
 export default function EntryForm() {
   const { push } = useRouter();
