--- src/components/tabs/history-tab.tsx
@@ -1,3 +1,0 @@
-export default function HistoryTab() {
-  return <div>Feature coming soon</div>;
-}
