+++ src/pages/dashboard.tsx
@@ -1,13 +1,103 @@
 import Link from "next/link";
-import Header from "@/components/common/header";
-import Loader from "@/components/common/loader";
+import Header from "@/components/ui/header";
+import Loader from "@/components/ui/loader";
 import WorkoutCard from "@/components/cards/workout-card";
-import DashboardTabs from "@/components/tabs/dashboard-tabs";
+import SearchBar from "@/components/ui/search-bar";
+import ExerciseCard from "@/components/cards/exercise-card";
+import TabComponent from "@/components/ui/tab";
 import { authorize } from "@/utils/authorize";
+import { Tab } from "@headlessui/react";
 import { useSession } from "next-auth/react";
-import { trpc } from "@/utils/trpc";
+import { InferQueryOutput, trpc } from "@/utils/trpc";
+import { useState } from "react";
+import { useSearch } from "@/utils/use-search";
 
 export { authorize as getServerSideProps };
+
+const Tabs = () => {
+  const [workoutQuery, setWorkoutQuery] = useState("");
+  const [exerciseQuery, setExerciseQuery] = useState("");
+
+  const workouts = trpc.useQuery(["workout.get-all"]);
+  const exercises = trpc.useQuery(["exercise.get-all"]);
+
+  const filteredWorkoutData = useSearch(
+    workoutQuery,
+    workouts.data
+  ) as InferQueryOutput<"workout.get-all">;
+  const filteredExerciseData = useSearch(
+    exerciseQuery,
+    exercises.data
+  ) as InferQueryOutput<"exercise.get-all">;
+
+  if (workouts.isLoading || exercises.isLoading) return <Loader />;
+
+  return (
+    <Tab.Group>
+      <Tab.List className='flex overflow-x-auto scrollbar-hide'>
+        <TabComponent>Workouts</TabComponent>
+        <TabComponent>Exercises</TabComponent>
+        <TabComponent>History</TabComponent>
+      </Tab.List>
+
+      <Tab.Panels>
+        <Tab.Panel className='space-y-6'>
+          <SearchBar
+            type='workout'
+            query={workoutQuery}
+            setQuery={setWorkoutQuery}
+          />
+
+          <Link className='link' href='/entry/create?option=workout'>
+            + Create workout
+          </Link>
+
+          {filteredWorkoutData && filteredWorkoutData?.length <= 0 && (
+            <p className='p-4 bg-zinc-900 border border-zinc-500 rounded-md'>
+              No workouts found.
+            </p>
+          )}
+
+          <div className='grid grid-cols-1 gap-8 sm:grid-cols-2 md:grid-cols-3'>
+            {filteredWorkoutData &&
+              filteredWorkoutData.map(workout => (
+                <WorkoutCard key={workout.id} workout={workout} />
+              ))}
+          </div>
+        </Tab.Panel>
+
+        <Tab.Panel className='space-y-6'>
+          <SearchBar
+            type='exercise'
+            query={exerciseQuery}
+            setQuery={setExerciseQuery}
+          />
+
+          <Link className='link' href='/exercise/create'>
+            + Create exercise
+          </Link>
+
+          {filteredExerciseData && filteredExerciseData?.length <= 0 && (
+            <p className='p-4 bg-zinc-900 border border-zinc-500 rounded-md'>
+              No exercises found.
+            </p>
+          )}
+
+          <div className='grid grid-cols-1 gap-8 sm:grid-cols-2 md:grid-cols-3'>
+            {filteredExerciseData &&
+              filteredExerciseData?.map(exercise => (
+                <ExerciseCard key={exercise.id} exercise={exercise as any} />
+              ))}
+          </div>
+        </Tab.Panel>
+
+        <Tab.Panel>
+          <p>Coming soon</p>
+        </Tab.Panel>
+      </Tab.Panels>
+    </Tab.Group>
+  );
+};
 
 export default function Dashboard() {
   const { data: session } = useSession();
@@ -49,7 +139,7 @@ export default function Dashboard() {
 
         <h2 className='text-lg font-bold sm:text-xl'>Explore</h2>
 
-        <DashboardTabs />
+        <Tabs />
       </main>
     </>
   );
