+++ src/pages/workout/[id]/edit.tsx
@@ -1,28 +1,25 @@
-import Header from "@/components/common/header";
-import Loader from "@/components/common/loader";
+import Header from "@/components/ui/header";
+import Loader from "@/components/ui/loader";
 import { authorize } from "@/utils/authorize";
 import { useRouter } from "next/router";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { convertKgToLbs } from "@/utils/kg-to-lbs";
-import { useProfileStore } from "@/utils/stores";
 import { trpc } from "@/utils/trpc";
 import toast from "react-hot-toast";
 import {
   updateWorkoutValidator,
   UpdateWorkoutValidator
-} from "@/utils/validators/update-workout";
+} from "@/server/shared/update-workout";
 
 export { authorize as getServerSideProps };
 
-interface Props {
+type Props = {
   workoutId: string;
-}
+};
 
 const EditWorkoutForm = ({ workoutId }: Props) => {
   const { push } = useRouter();
   const ctx = trpc.useContext();
-  const { weightUnit } = useProfileStore();
 
   const { data: workout, isLoading } = trpc.useQuery([
     "workout.get-by-id",
