+++ src/server/auth.ts
@@ -1,0 +1,8 @@
+import { GetServerSidePropsContext } from "next";
+import { unstable_getServerSession } from "next-auth";
+import { authOptions } from "@/pages/api/auth/[...nextauth]";
+
+export const getUpsurgeAuth = (ctx: {
+  req: GetServerSidePropsContext["req"];
+  res: GetServerSidePropsContext["res"];
+}) => unstable_getServerSession(ctx.req, ctx.res, authOptions);
