+++ src/server/context.ts
@@ -1,14 +1,10 @@
 import * as trpc from "@trpc/server";
 import * as trpcNext from "@trpc/server/adapters/next";
-import { unstable_getServerSession } from "next-auth";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
 import { prisma } from "./db";
+import { getUpsurgeAuth } from "./auth";
 
-export const createContext = async ({
-  req,
-  res
-}: trpcNext.CreateNextContextOptions) => {
-  const session = await unstable_getServerSession(req, res, authOptions);
+export const createContext = async (ctx: trpcNext.CreateNextContextOptions) => {
+  const session = await getUpsurgeAuth(ctx);
   return { session, prisma };
 };
