+++ tailwind.config.js
@@ -6,17 +6,7 @@ module.exports = {
       center: true,
       padding: "2rem"
     },
-    extend: {
-      animation: {
-        scale: "scale 150ms linear"
-      },
-      keyframes: {
-        scale: {
-          from: { opacity: 0, transform: "scale(0.75)" },
-          to: { opacity: 1, transform: "scale(1)" }
-        }
-      }
-    }
+    extend: {}
   },
   plugins: [require("@tailwindcss/line-clamp")]
 };
